requirements.txt
tensorflow
opencv-python
numpy
matplotlib
flask
pycryptodome
scikit-learn